koki-to-kube of different Deployment versions
