+++ examples/deployment/example1.short.yaml
@@ -11,4 +11,4 @@ deployment:
   replicas: 3
   selector:
     app: example
-  version: extensions/v1beta1
+  version: apps/v1beta1
